Add missing dependencies
We can't import keys without PGP suite installed. Adding. Add another check to install Docker automated only on Debian based distributions.

+++ compile.sh
@@ -42,11 +42,11 @@ if [[ "$1" == vagrant && -z "$(which vagrant)" ]]; then
 	sudo apt-get install -y vagrant virtualbox
 fi
 
-# Install Docker if not there but wanted
-if [[ "$1" == docker && -z "$(which docker)" ]]; then
+# Install Docker if not there but wanted. We cover only Debian based distro install. Else, manual Docker install is needed
+if [[ "$1" == docker && -f /etc/debian_version && -z "$(which docker)" ]]; then
 	display_alert "Docker not installed." "Installing" "Info"
 	echo "deb https://download.docker.com/linux/$(lsb_release -is | awk '{print tolower($0)}') $(lsb_release -cs) edge" > /etc/apt/sources.list.d/docker.list
-	[[ ! $(which curl) ]] && apt-get update;apt-get install -y -qq --no-install-recommends curl
+	[[ ! $(which curl) ]] && apt-get update;apt-get install -y -qq --no-install-recommends curl gnupg
 	curl -fsSL "https://download.docker.com/linux/$(lsb_release -is | awk '{print tolower($0)}')/gpg" | apt-key add -qq - > /dev/null 2>&1
 	export DEBIAN_FRONTEND=noninteractive
 	apt-get update
